requirements.txt
Flask==0.11.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
pytest-flask==0.10.0
python-decouple==3.0
uWSGI==2.0.14
Flask-Collect==1.3.2
ansible==2.2.0.0
Flask-Migrate==2.0.1
psycopg2==2.6.2
Flask-WTF==0.13.1
